Abstract our partial header serializer
This moves the partial header serializer out so it can be tested, it
also upgrades the test so that it doesn't use the fake strategy.
