+++ ironfish/src/mining/__fixtures__/director.old.test.slow.ts.fixture
